requirements.txt
prettyconf
google-generativeai
rich
